change "double" to "numeric"

+++ tests/testthat/test-my_rf_cv.R
@@ -1,3 +1,3 @@
 test_that("my_rf works", {
-  expect_is(my_rf_cv(5), "double")
+  expect_is(my_rf_cv(5), "numeric")
 })
